requirements-api.txt
# requirements-api.txt — LevelUp-ELO V2.0 API (FastAPI)
# Instalar con: pip install -r requirements-api.txt
# Incluye todas las deps de V1 + FastAPI stack.

# ── V1 base ──────────────────────────────────────────────────────────────────
passlib[argon2]>=1.7,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
sympy>=1.12,<2.0
openai>=1.0.0,<2.0
anthropic>=0.40.0,<1.0
Pillow>=10.0,<11.0
openpyxl>=3.1,<4.0
supabase>=2.0,<3.0
matplotlib>=3.8,<4.0

# ── FastAPI stack ─────────────────────────────────────────────────────────────
fastapi>=0.111,<1.0
uvicorn[standard]>=0.29,<1.0
python-jose[cryptography]>=3.3,<4.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
slowapi>=0.1.9,<1.0
python-multipart>=0.0.9,<1.0

# ── Tests FastAPI ─────────────────────────────────────────────────────────────
httpx>=0.27,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=5.0,<6.0
